pyproject.toml
[project]
name = "morph-example-0.3.0"
version = "0.1.0"
description = ""
authors = [{ name = "Keita Mitsuhashi", email = "keita.mitsuhashi.83@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "morph-data==0.3.0rc1"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
